feat: created sqlite driver with init schema
