+++ backend/src/repositories/index.ts
@@ -1,4 +1,4 @@
-export { getDatabase } from "./sqlite.js";
+export { getDatabase } from "../drivers/sqlite.js";
 export { metadataRepository } from "./metadata.js";
-export { transactionResultsRepository } from "./transactionResults.js";
 export { nodesRepository } from "./nodes.js";
+export { transactionResultsRepository } from "./transactionResults.js";
