+++ backend/src/repositories/metadata.ts
@@ -1,25 +1,15 @@
 import { OffchainMetadata } from "@autonomys/auto-drive";
-import { getDatabase } from "./sqlite.js";
+import { getDatabase } from "../drivers/sqlite.js";
 
 export interface MetadataEntry<T extends string | OffchainMetadata = string> {
   cid: string;
   metadata: T;
 }
 
-const initTable = async () => {
-  const db = await getDatabase();
-
-  db.exec(
-    "CREATE TABLE IF NOT EXISTS metadata (cid TEXT PRIMARY KEY, metadata TEXT)",
-  );
-
-  return db;
-};
-
 const getMetadata = async (
-  cid: string,
+  cid: string
 ): Promise<MetadataEntry<OffchainMetadata> | undefined> => {
-  const db = await initTable();
+  const db = await getDatabase();
 
   return db
     .get<MetadataEntry>("SELECT * FROM metadata WHERE cid = ?", cid)
@@ -31,22 +21,24 @@ const getMetadata = async (
 };
 
 const setMetadata = async (cid: string, metadata: OffchainMetadata) => {
-  const db = await initTable();
+  const db = await getDatabase();
 
   return db.run(
     "INSERT OR REPLACE INTO metadata (cid, metadata) VALUES (?, ?)",
     cid,
-    JSON.stringify(metadata),
+    JSON.stringify(metadata)
   );
 };
 
 const searchMetadataByCID = async (cid: string, limit: number | undefined) => {
-  const db = await initTable();
+  const db = await getDatabase();
 
   return db
-    .all<
-      { cid: string }[]
-    >("SELECT cid FROM metadata WHERE cid LIKE ? LIMIT ?", `%${cid}%`, limit)
+    .all<{ cid: string }[]>(
+      "SELECT cid FROM metadata WHERE cid LIKE ? LIMIT ?",
+      `%${cid}%`,
+      limit
+    )
     .then((entries) => {
       return entries.map((entry) => {
         return {
@@ -58,7 +50,7 @@ const searchMetadataByCID = async (cid: string, limit: number | undefined) => {
 
 /// TODO: pagination and filtering
 export const getAllMetadata = async () => {
-  const db = await initTable();
+  const db = await getDatabase();
 
   return db.all<MetadataEntry[]>("SELECT * FROM metadata").then((entries) => {
     return entries.map((entry) => {
